@@ -1,8 +1,12 @@
+# frozen_string_literal: true
+
 module Stator
+
   MAJOR       = 0
   MINOR       = 3
-  PATCH       = 0
+  PATCH       = 1
   PRERELEASE  = nil
 
-  VERSION = [MAJOR, MINOR, PATCH, PRERELEASE].compact.join('.')
+  VERSION = [MAJOR, MINOR, PATCH, PRERELEASE].compact.join(".")
+
 end
